+++ src/commands/mod.rs
@@ -1,5 +1,6 @@
 pub mod config;
 pub mod gc;
+pub mod hook;
 pub mod init;
 pub mod projects;
 pub mod simple;